requirements-linux.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
base58==2.1.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
construct==2.10.68
construct-typing==0.6.2
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jedi==0.19.0
jupyter==1.0.0
jupyter-client==8.3.0
jupyter-console==6.6.0
jupyter-core==5.3.0
jupyter-server==2.7.0
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lxml==4.9.3
Markdown==3.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.7
notebook==7.0.0
notebook-shim==0.2.3
numpy==2.1.3
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==25.1.1
platformdirs==3.10.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.16.1
pyparsing==3.1.1
pyrsistent==0.20.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytoolconfig==1.3.1
pytz==2024.2
# Removed pywin32 for Linux compatibility
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.0
requests==2.31.0
scikit-learn==1.6.1
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.2
tensorflow==2.19.0
tensorflow-estimator==2.19.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.10.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==3.0.1
wrapt==1.15.0
zipp==3.17.0

# Discord.py for Discord bot functionality
discord.py==2.3.2

# Google Drive Integration
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Additional utilities
pydrive2==1.20.0

# Voice support for Discord (optional but recommended)
PyNaCl==1.5.0

# Technical Analysis Library (TA-Lib)
# Note: ta-lib requires system-level installation in Docker
# Added in Dockerfile: apt-get install -y libta-lib-dev
